Refactor DOMTreeView component: move to common directory, remove unused CSS, and enhance styling for improved organization and usability
