+++ src/components/common/DOMTreeView.tsx
@@ -1,10 +1,9 @@
 // DOMTreeView.tsx
 import { ChevronDown, ChevronRight } from 'lucide-react';
 import React, { useState } from 'react';
-import { ElementInfo } from '../types/domSelection';
-import './DOMTreeView.css';
-import { Tooltip } from './common/Tooltip';
-import { formatElementTag } from './utils/htmlTagFormatter';
+import { ElementInfo } from '../../types/domSelection';
+import { formatElementTag } from '../../utils/htmlTagFormatter';
+import { Tooltip } from './Tooltip';
 
 interface Props {
   /** The element information to display in the tree view */
