+++ src/styles/common.css
@@ -10,6 +10,11 @@
   --focus-ring-color: #3b82f6;
   --shadow-sm: 0 1px 2px 0 rgba(0, 0, 0, 0.05);
   --shadow-md: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -2px rgba(0, 0, 0, 0.05);
+  --tree-indent: 0.75rem;
+  --tree-line-color: #e5e7eb;
+  --tree-chevron-color: #6b7280;
+  --tree-tag-color: #2563eb;
+  --monospace-font: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, monospace;
 }
 
 *,
@@ -140,4 +145,73 @@ button:hover {
 /* Extend style: Success tooltip */
 .tooltip.success {
   background-color: rgba(34, 197, 94, 0.9);
+}
+
+/* Tree view styles */
+.tree-container {
+  padding: 8px 0;
+  font-family: var(--monospace-font);
+  font-size: 13px;
+  line-height: 1.5;
+  max-width: 100%;
+}
+
+.tree-node {
+  margin-left: var(--tree-indent);
+  max-width: calc(100% - calc(var(--tree-indent) * 2));
+}
+
+.tree-node-content {
+  display: flex;
+  align-items: flex-start;
+  gap: 0.25rem;
+  padding: 0.25rem 0;
+  position: relative;
+}
+
+.tree-chevron {
+  flex-shrink: 0;
+  width: 16px;
+  height: 16px;
+  cursor: pointer;
+  color: var(--tree-chevron-color);
+  transition: transform 0.2s ease;
+}
+
+.tree-chevron-placeholder {
+  width: 16px;
+  height: 16px;
+}
+
+.tree-tag-container {
+  overflow: hidden;
+  min-width: 0;
+}
+
+.tree-tag {
+  color: var(--tree-tag-color);
+  cursor: pointer;
+  overflow-wrap: break-word;
+  word-break: break-word;
+  display: inline-block;
+  max-width: 100%;
+}
+
+.tree-tag:hover {
+  text-decoration: underline;
+}
+
+.tree-children {
+  position: relative;
+  padding-left: var(--tree-indent);
+}
+
+.tree-children::before {
+  content: '';
+  position: absolute;
+  left: 0;
+  top: 0;
+  bottom: 0;
+  width: 1px;
+  background-color: var(--tree-line-color);
 }
